ee/intelligent_search/requirements.txt
# General utils
pydantic==2.3.0
requests==2.31.0
python-decouple==3.8
certifi==2023.7.22

# AWS utils
awscli==1.29.53

# ML modules
# torch==2.0.1
fairscale==0.4.13
sentencepiece==0.1.99

# Serving modules
fastapi==0.103.1
httpx==0.25.0
apscheduler==3.10.4
uvicorn==0.23.2

# Observability modules
traceloop-sdk==0.0.37

# Test
pytest==7.4.2
